env_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.3.1
alembic=1.15.2=pypi_0
blas=1.0=mkl
bottle=0.13.2=py310haa95532_0
bottleneck=1.4.2=py310hc99e966_0
brotli=1.0.9=h827c3e9_9
brotli-bin=1.0.9=h827c3e9_9
brotli-python=1.0.9=py310h5da7b33_9
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
certifi=2025.4.26=py310haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
colorlog=6.9.0=pypi_0
contourpy=1.3.1=py310h214f63a_0
cycler=0.11.0=pyhd3eb1b0_0
deprecated=1.2.13=py310haa95532_0
filelock=3.13.1=py310haa95532_0
fonttools=4.55.3=py310h827c3e9_0
freetype=2.13.3=h0620614_0
fsspec=2024.12.0=py310haa95532_0
gmp=6.3.0=h537511b_0
gmpy2=2.2.1=py310h827c3e9_0
greenlet=3.1.1=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
idna=3.7=py310haa95532_0
importlib-metadata=8.5.0=py310haa95532_0
importlib_resources=6.4.0=py310haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.6=py310haa95532_0
joblib=1.4.2=pypi_0
jpeg=9e=h827c3e9_3
kiwisolver=1.4.8=py310h5da7b33_0
krb5=1.20.1=h5b6d351_0
lcms2=2.16=hb4a4139_0
lerc=4.0.0=h5da7b33_0
libabseil=20240116.2=cxx17_h5da7b33_0
libbrotlicommon=1.0.9=h827c3e9_9
libbrotlidec=1.0.9=h827c3e9_9
libbrotlienc=1.0.9=h827c3e9_9
libdeflate=1.22=h5bf469e_0
libffi=3.4.4=hd77b12b_1
libpng=1.6.39=h8cc25b3_0
libpq=17.4=h70ee33d_0
libprotobuf=4.25.3=hf2fb9eb_0
libtiff=4.5.1=h44ae7cf_1
libtorch=2.5.1=cpu_mkl_h1aab2ed_102
libuv=1.48.0=h827c3e9_0
libwebp-base=1.3.2=h3d04722_1
lz4-c=1.9.4=h2bbff1b_1
mako=1.3.9=pypi_0
markupsafe=3.0.2=py310h827c3e9_0
matplotlib=3.10.0=py310haa95532_0
matplotlib-base=3.10.0=py310he19b0ae_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py310h827c3e9_2
mkl_fft=1.3.11=py310h827c3e9_0
mkl_random=1.2.8=py310hc64d2fc_0
mpc=1.3.1=h827c3e9_0
mpfr=4.2.1=h56c3642_0
mpmath=1.3.0=py310haa95532_0
networkx=3.4.2=py310haa95532_0
numexpr=2.10.1=py310h4cd664f_0
numpy=1.26.4=py310h055cbcc_0
numpy-base=1.26.4=py310h65a83cf_0
openjpeg=2.5.2=hae555c5_0
openssl=3.1.0=hcfcfb64_3
opentelemetry-api=1.30.0=py310haa95532_0
optuna=4.2.1=pypi_0
optuna-dashboard=0.18.0=pypi_0
packaging=24.2=py310haa95532_0
pandas=2.2.3=py310h5da7b33_0
pcre2=10.42=h0ff8eda_1
pillow=11.1.0=py310h096bfcc_0
pip=25.0=py310haa95532_0
platformdirs=3.10.0=py310haa95532_0
plotly=5.24.1=py310h9909e9c_1
pooch=1.8.2=py310haa95532_0
pyparsing=3.2.0=py310haa95532_0
pyqt=6.7.1=py310h5da7b33_0
pyqt6-sip=13.9.1=py310h827c3e9_0
pysocks=1.7.1=py310haa95532_0
python=3.10.16=h4607a30_1
python-dateutil=2.9.0post0=py310haa95532_2
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.5.1=cpu_mkl_py310hc9ee3b7_102
pytz=2024.1=py310haa95532_0
pyyaml=6.0.2=pypi_0
qtbase=6.7.3=h0804d20_0
qtdeclarative=6.7.3=h5da7b33_0
qtsvg=6.7.3=hf2fb9eb_0
qttools=6.7.3=h0de5f00_0
qtwebchannel=6.7.3=h5da7b33_0
qtwebsockets=6.7.3=h5da7b33_0
requests=2.32.3=py310haa95532_1
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=py310h9d85e7c_1
seaborn=0.13.2=py310haa95532_2
setuptools=72.1.0=py310haa95532_0
sip=6.10.0=py310h5da7b33_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h8cc25b3_2
sqlalchemy=2.0.40=pypi_0
sqlite=3.45.3=h2bbff1b_0
swig=4.3.0=hf2fb9eb_0
sympy=1.13.3=py310haa95532_1
tbb=2021.8.0=h59b6b97_0
tenacity=9.0.0=py310haa95532_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h0416ee5_0
tomli=2.0.1=py310haa95532_0
tornado=6.4.2=py310h827c3e9_0
tqdm=4.67.1=pypi_0
typing_extensions=4.12.2=py310haa95532_0
tzdata=2025a=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_1
unicodedata2=15.1.0=py310h827c3e9_1
urllib3=2.3.0=py310haa95532_0
vc=14.42=haa95532_4
vs2015_runtime=14.42.34433=he0abc0d_4
wheel=0.45.1=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
wrapt=1.17.0=py310h827c3e9_0
xz=5.6.4=h4754444_1
zipp=3.21.0=py310haa95532_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
